Update ICM456_StillV0.2_2_15625
ICM is completely still for ~1min. Time is micros()/100000 into seconds. 
The problem with the mapping might be that the package expects the start to be 0 seconds and not 6 or 18seconds (which is why it gets so extreme so early, meaning the rest of the map stays extreme - chain reaction).